Increse timeout to 390 minutes Kci and Rci debug and release
These 4 jobs have been struggling to finish because most of them are timing out.
The builds that finish are near the 360 minutes threshold, so some of them fail and some others finish.

See the current build trend of realted jobs (May 29th, 2025):
- https://build.ros2.org/view/Rci/job/Rci__nightly-debug_ubuntu_noble_amd64/buildTimeTrend
- https://build.ros2.org/view/Rci/job/Rci__nightly-release_ubuntu_noble_amd64/buildTimeTrend
- https://build.ros2.org/view/Kci/job/Kci__nightly-debug_ubuntu_noble_amd64/buildTimeTrend
- https://build.ros2.org/view/Kci/job/Kci__nightly-release_ubuntu_noble_amd64/buildTimeTrend

A reason of these failures may come from the amount of test regressions that are happening on the buildfarm right now,
as most of them need to be re-tested multiple times with the `--retest-until-pass 2` flag.

Signed-off-by: Crola1702 <cristobal.arroyo@ekumenlabs.com>