Use nightly-release as underlay for extra-rmw-release
At present, the only difference between `nightly-release` and
`extra-rmw-release` jobs is the latter's inclusion of
`rmw_fastrtps_dynamic_cpp`. Rather than re-build and re-test EVERYTHING
in `extra-rmw-release`, we should re-use the `nightly-release` and only
re-build what we might need.

The result is that we'll re-use roughly half of the packages which are
identical between these two jobs, and won't need to build or test them
in this job for no reason. Because such a significant amount of this
job's runtime is spent running cross-vendor tests I don't anticipate it
will have a proportional impact on time savings, but it should be
significant.

This change should be just as "safe" as the original approach because it
re-builds everything that might be downstream of the additional
packages. In the future, we may be able to scope this further to omit
downstream packages which we know won't be affected by the additional
RMW's presence.

Signed-off-by: Scott K Logan <logans@cottsay.net>